Move custom bounds test to ui test
